Release AI-ModelZoo-2.0.1:
   - Disclosed some ST object detection models: `st_yolo_lc_v1` and `ssd_mobilenet_v2_fpnlite` with various resolutions.
   - Disclosed some ST image classification models: `st_efficientnet_lc_v1`, `st_fdmobilenet_v1`, `st_resnet_8_hybrid_v1`, and `st_resnet_8_hybrid_v2` in different resolutions.
   - Fixed minor bugs and updated README documentation.

+++ object_detection/src/quantization/quantize.py
@@ -62,7 +62,7 @@ def tfite_ptq_quantizer(
 
     def fake_data_gen():
         for _ in tqdm.tqdm(range(num_fake_images)):
-            image = tf.random.uniform(image_shape, minval=0, maxval=256, dtype=tf.int8)
+            image = tf.random.uniform(image_shape, minval=0, maxval=256, dtype=tf.int32)
             image = tf.cast(image, tf.float32)
             
             # Rescale the image
